fix:修复登录时mid int太小装不下

+++ shared/auth/src/commonMain/kotlin/top/suzhelan/bili/shared/auth/config/LoginMapper.kt
@@ -91,7 +91,7 @@ object LoginMapper {
         _isLoginState.value = false
     }
 
-    fun getMid(): Int {
+    fun getMid(): Long {
         return getUniversalLoginInfo().tokenInfo.mid
     }
